requirements.txt
amqp==1.4.9
anyjson==0.3.3
arrow==0.10.0
beautifulsoup4==4.4.1
billiard==3.3.0.23
boto3==1.3.1
botocore==1.4.93
celery==3.1.23
certifi==2017.7.27.1
chardet==3.0.4
contextlib2==0.5.5
coverage==4.2
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.7.11
django-cors-headers==1.1.0
django-dotenv==1.4.1
django-extensions==1.6.1
django-filter==0.7
django-grappelli==2.8.1
django-guardian==1.4.4
django-haystack==2.4.1
django-import-export==0.4.5
django-reversion==1.8.6
django-user-agents==0.3.0
djangorestframework==2.3.14
docutils==0.14
drf-nested-routers==0.11.1
enum34==1.1.6
et-xmlfile==1.0.1
Faker==0.8.7
flake8==3.0.4
fuzzywuzzy==0.10.0
geopy==1.11.0
googlemaps==2.4.4
idna==2.6
jdcal==1.3
jmespath==0.9.3
kombu==3.0.37
mccabe==0.5.3
oauthlib==2.0.6
odfpy==1.3.5
openpyxl==2.4.9
Pillow==2.7.0
psycopg2==2.7.3.2
pycodestyle==2.0.0
pyflakes==1.2.3
PyJWT==1.5.3
pypng==0.0.18
PyQRCode==1.1.1
python-dateutil==2.6.1
python-geohash==0.8.5
python-Levenshtein==0.12.0
python-social-auth==0.2.18
python3-openid==3.1.0
pytz==2017.3
PyYAML==3.11
raven==6.4.0
redis==2.10.5
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
suds-py3==1.3.2.0
tablib==0.12.1
text-unidecode==1.1
ua-parser==0.7.0
unicodecsv==0.14.1
urllib3==1.22
user-agents==1.0.1
Whoosh==2.7.4
xlrd==1.1.0
xlwt==1.3.0
gevent==1.2.2
